Feature #4056 Update poms, fix tests.

+++ globalpomutils-core/src/test/groovy/com/anrisoftware/globalpom/core/pointformat/PointFormatTest.groovy
@@ -1,31 +1,15 @@
-/*
- * Copyright 2016 Erwin Müller <erwin.mueller@deventm.org>
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
 package com.anrisoftware.globalpom.core.pointformat
 
 import static com.anrisoftware.globalpom.utils.TestUtils.*
-import groovy.transform.CompileStatic
 
 import java.awt.Point
 import java.awt.geom.Point2D
 import java.text.ParseException
 import java.text.ParsePosition
 
-import org.junit.Test
+import org.junit.jupiter.api.Test
 
-import com.anrisoftware.globalpom.core.pointformat.PointFormat
+import groovy.transform.CompileStatic
 
 /**
  * Test parsing and formatting of a point.
@@ -42,7 +26,7 @@ class PointFormatTest {
         def str = format.format new Point2D.Double(1, 2)
         assertStringContent "(1, 2)", str
 
-        str = format.format new Point2D.Double(1.1, 2.1)
+        str = format.format new Point2D.Double((double)1.1, (double)2.1)
         assertStringContent "(1.1, 2.1)", str
 
         str = format.format new Point(1, 2)
